// sanity/sanity.types.ts
export type JournalEntry = {
  _id: string
  _type: 'journalEntry'
  _createdAt: string
  _updatedAt: string
  _rev: string
  title?: string
  content?: (
    | {
        children?: {
          marks?: string[]
          text?: string
          _type: 'span'
          _key: string
        }[]
        style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote'
        listItem?: 'bullet' | 'number'
        markDefs?: {
          href?: string
          _type: 'link'
          _key: string
        }[]
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        media?: unknown
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        alt?: string
        caption?: string
        _type: 'image'
        _key: string
      }
  )[]
  mood?: 'very-sad' | 'sad' | 'neutral' | 'happy' | 'very-happy'
  userId?: string
  aiGeneratedCategory?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'category'
  }
  createdAt?: string
}
// ...
export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}

export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol